Refactor tvmaze_api -> api
