+++ medusa/indexers/config.py
@@ -7,7 +7,7 @@ import re
 from medusa.app import BASE_PYMEDUSA_URL
 from medusa.indexers.tmdb.tmdb import Tmdb
 from medusa.indexers.tvdbv2.tvdbv2_api import TVDBv2
-from medusa.indexers.tvmaze.tvmaze_api import TVmaze
+from medusa.indexers.tvmaze.api import TVmaze
 from medusa.session.core import MedusaSession
